+++ src/messages/MessagesBuilder.ts
@@ -7,16 +7,16 @@ import {
 import { Envelope, StepDefinitionPatternType } from '@cucumber/messages'
 
 import { extractStepTexts } from '../gherkin/extractStepTexts.js'
-import { buildStepDocuments } from '../step-documents/buildStepDocuments.js'
-import { StepDocument } from '../step-documents/types.js'
+import { buildSuggestions } from '../suggestions/buildSuggestions.js'
+import { Suggestion } from '../suggestions/types.js'
 
 export type MessagesBuilderResult = {
-  stepDocuments: readonly StepDocument[]
+  suggestions: readonly Suggestion[]
   expressions: readonly Expression[]
 }
 
 /**
- * Builds CucumberInfo from Cucumber Messages.
+ * Builds MessagesBuilderResult from Cucumber Messages.
  */
 export class MessagesBuilder {
   private readonly parameterTypeRegistry = new ParameterTypeRegistry()
@@ -65,7 +65,7 @@ export class MessagesBuilder {
 
   build(): MessagesBuilderResult {
     return {
-      stepDocuments: buildStepDocuments(this.stepTexts, this.expressions),
+      suggestions: buildSuggestions(this.parameterTypeRegistry, this.stepTexts, this.expressions),
       expressions: this.expressions,
     }
   }