+++ test/messages/MessagesBuilder.test.ts
@@ -5,7 +5,7 @@ import { pipeline as pipelineCb, Writable } from 'stream'
 import { promisify } from 'util'
 
 import { MessagesBuilderResult } from '../../src/messages/MessagesBuilder.js'
-import { StepDocument } from '../../src/step-documents/types.js'
+import { Suggestion } from '../../src/suggestions/types.js'
 import { MessagesBuilderStream } from './MessagesBuilderStream.js'
 
 const pipeline = promisify(pipelineCb)
@@ -63,46 +63,43 @@ describe('MessagesBuilder', () => {
         },
       })
     )
-    const expectedStepDocuments: Partial<StepDocument>[] = [
+    const expectedSuggestions: Suggestion[] = [
       {
         segments: ['I select the ', ['2nd'], ' snippet'],
-        suggestion: 'I select the {ordinal} snippet',
+        label: 'I select the {ordinal} snippet',
       },
       {
         segments: [
           'I type ',
           ['"I have ${1|11,17,23|} cukes on my ${2|belly,table,tree|}"', '"cukes"'],
         ],
-        suggestion: 'I type {string}',
+        label: 'I type {string}',
+      },
+      {
+        segments: ['the following Gherkin step texts exist:'],
+        label: 'the following Gherkin step texts exist:',
+      },
+      {
+        segments: ['the following Step Definitions exist:'],
+        label: 'the following Step Definitions exist:',
       },
       {
         segments: [
           'the LSP snippet should be ',
           ['"I have ${1|11,17,23|} cukes on my ${2|belly,table,tree|}"', '"cukes"'],
         ],
-        suggestion: 'the LSP snippet should be {string}',
-      },
-      {
-        segments: ['the following Gherkin step texts exist:'],
-        suggestion: 'the following Gherkin step texts exist:',
-      },
-      {
-        segments: ['the following Step Definitions exist:'],
-        suggestion: 'the following Step Definitions exist:',
+        label: 'the LSP snippet should be {string}',
       },
       {
         segments: ['the suggestions should be empty'],
-        suggestion: 'the suggestions should be empty',
+        label: 'the suggestions should be empty',
       },
       {
         segments: ['the suggestions should be:'],
-        suggestion: 'the suggestions should be:',
+        label: 'the suggestions should be:',
       },
     ]
-    assert.deepStrictEqual(
-      result!.stepDocuments.map((d) => ({ segments: d.segments, suggestion: d.suggestion })),
-      expectedStepDocuments
-    )
+    assert.deepStrictEqual(result!.suggestions, expectedSuggestions)
 
     const expectedExpressionSources = [
       'the following Gherkin step texts exist:',